Moved cache code to draw2dbase
